Refactoring FollowCommentsUiState.
